Renamed deploy_messages -> sd_messages.
