+++ simple_deploy/management/commands/sd_messages.py
